+++ tests/tests.hs
@@ -5,6 +5,7 @@
 module Main where
 
 import           Control.Applicative
+import           Control.Exception
 import           Control.Monad
 import           Control.Monad.Reader
 import           Data.Aeson
@@ -13,6 +14,7 @@ import           Data.List                       (nub)
 import           Data.List.NonEmpty              (NonEmpty (..))
 import qualified Data.List.NonEmpty              as NE
 import qualified Data.Map.Strict                 as M
+import           Data.Monoid
 import           Data.Text                       (Text)
 import qualified Data.Text                       as T
 import           Data.Time.Calendar              (Day (..))
@@ -25,7 +27,7 @@ import           Network.HTTP.Client
 import qualified Network.HTTP.Types.Status       as NHTS
 import           Prelude                         hiding (filter)
 import           Test.Hspec
-import           Test.QuickCheck.Property.Monoid
+import           Test.QuickCheck.Property.Monoid (prop_Monoid, eq, T(..))
 
 import           Test.Hspec.QuickCheck           (prop)
 import           Test.QuickCheck
@@ -46,7 +48,7 @@ validateStatus resp expected =
   `shouldBe` (expected :: Int)
 
 createExampleIndex :: BH IO Reply
-createExampleIndex = createIndex defaultIndexSettings testIndex
+createExampleIndex = createIndex (IndexSettings (ShardCount 1) (ReplicaCount 0)) testIndex
 deleteExampleIndex :: BH IO Reply
 deleteExampleIndex = deleteIndex testIndex
 
@@ -147,32 +149,42 @@ otherTweet = Tweet { user     = "notmyapp"
                    , age      = 1000
                    , location = Location 40.12 (-71.34) }
 
-insertData :: BH IO ()
-insertData = do
+resetIndex :: BH IO ()
+resetIndex = do
   _ <- deleteExampleIndex
   _ <- createExampleIndex
   _ <- putMapping testIndex testMapping TweetMapping
-  _ <- indexDocument testIndex testMapping exampleTweet (DocId "1")
-  _ <- refreshIndex testIndex
   return ()
+
+insertData :: BH IO Reply
+insertData = do
+  resetIndex
+  insertData' defaultIndexDocumentSettings
+
+insertData' :: IndexDocumentSettings -> BH IO Reply
+insertData' ids = do
+  r <- indexDocument testIndex testMapping ids exampleTweet (DocId "1")
+  _ <- refreshIndex testIndex
+  return r
 
 insertOther :: BH IO ()
 insertOther = do
-  _ <- indexDocument testIndex testMapping otherTweet (DocId "2")
+  _ <- indexDocument testIndex testMapping defaultIndexDocumentSettings otherTweet (DocId "2")
   _ <- refreshIndex testIndex
   return ()
 
 searchTweet :: Search -> BH IO (Either String Tweet)
 searchTweet search = do
-  reply <- searchByIndex testIndex search
-  let result = eitherDecode (responseBody reply) :: Either String (SearchResult Tweet)
+  result <- searchTweets search
   let myTweet = fmap (hitSource . head . hits . searchHits) result
   return myTweet
 
+searchTweets :: Search -> BH IO (Either String (SearchResult Tweet))
+searchTweets search = eitherDecode . responseBody <$> searchByIndex testIndex search
+
 searchExpectNoResults :: Search -> BH IO ()
 searchExpectNoResults search = do
-  reply <- searchByIndex testIndex search
-  let result = eitherDecode (responseBody reply) :: Either String (SearchResult Tweet)
+  result <- searchTweets search
   let emptyHits = fmap (hits . searchHits) result
   liftIO $
     emptyHits `shouldBe` Right []
@@ -204,8 +216,7 @@ searchTermsAggHint hints = do
 
 searchTweetHighlight :: Search -> BH IO (Either String (Maybe HitHighlight))
 searchTweetHighlight search = do
-  reply <- searchByIndex testIndex search
-  let result = eitherDecode (responseBody reply) :: Either String (SearchResult Tweet)
+  result <- searchTweets search
   let myHighlight = fmap (hitHighlight . head . hits . searchHits) result
   return myHighlight
 
@@ -268,6 +279,9 @@ instance Arbitrary a => Arbitrary (SearchHits a) where
     hs <- arbitrary
     return $ SearchHits tot score hs
 
+getSource :: EsResult a -> Maybe a
+getSource = fmap _source . foundResult
+
 main :: IO ()
 main = hspec $ do
 
@@ -288,8 +302,40 @@ main = hspec $ do
       docInserted <- getDocument testIndex testMapping (DocId "1")
       let newTweet = eitherDecode
                      (responseBody docInserted) :: Either String (EsResult Tweet)
-      liftIO $ (fmap _source newTweet `shouldBe` Right exampleTweet)
+      liftIO $ (fmap getSource newTweet `shouldBe` Right (Just exampleTweet))
 
+    it "produces a parseable result when looking up a bogus document" $ withTestEnv $ do
+      doc <- getDocument testIndex testMapping  (DocId "bogus")
+      let noTweet = eitherDecode
+                    (responseBody doc) :: Either String (EsResult Tweet)
+      liftIO $ fmap foundResult noTweet `shouldBe` Right Nothing
+
+    it "can use optimistic concurrency control" $ withTestEnv $ do
+      let ev = ExternalDocVersion minBound
+      let cfg = defaultIndexDocumentSettings { idsVersionControl = ExternalGT ev }
+      resetIndex
+      res <- insertData' cfg
+      liftIO $ isCreated res `shouldBe` True
+      res' <- insertData' cfg
+      liftIO $ isVersionConflict res' `shouldBe` True
+
+  describe "template API" $ do
+    it "can create a template" $ withTestEnv $ do
+      let idxTpl = IndexTemplate (TemplatePattern "tweet-*") (Just (IndexSettings (ShardCount 1) (ReplicaCount 1))) [toJSON TweetMapping]
+      resp <- putTemplate idxTpl (TemplateName "tweet-tpl")
+      liftIO $ validateStatus resp 200
+
+    it "can detect if a template exists" $ withTestEnv $ do
+      exists <- templateExists (TemplateName "tweet-tpl")
+      liftIO $ exists `shouldBe` True
+
+    it "can delete a template" $ withTestEnv $ do
+      resp <- deleteTemplate (TemplateName "tweet-tpl")
+      liftIO $ validateStatus resp 200
+
+    it "can detect if a template doesn't exist" $ withTestEnv $ do
+      exists <- templateExists (TemplateName "tweet-tpl")
+      liftIO $ exists `shouldBe` False
 
   describe "bulk API" $ do
     it "inserts all documents we request" $ withTestEnv $ do
@@ -308,8 +354,8 @@ main = hspec $ do
       let maybeFirst  = eitherDecode $ responseBody fDoc :: Either String (EsResult BulkTest)
       let maybeSecond = eitherDecode $ responseBody sDoc :: Either String (EsResult BulkTest)
       liftIO $ do
-        fmap _source maybeFirst `shouldBe` Right firstTest
-        fmap _source maybeSecond `shouldBe` Right secondTest
+        fmap getSource maybeFirst `shouldBe` Right (Just firstTest)
+        fmap getSource maybeSecond `shouldBe` Right (Just secondTest)
 
 
   describe "query API" $ do
@@ -341,8 +387,8 @@ main = hspec $ do
 
     it "returns document for multi-match query" $ withTestEnv $ do
       _ <- insertData
-      let fields = [FieldName "user", FieldName "message"]
-      let query = QueryMultiMatchQuery $ mkMultiMatchQuery fields (QueryString "bitemyapp")
+      let flds = [FieldName "user", FieldName "message"]
+      let query = QueryMultiMatchQuery $ mkMultiMatchQuery flds (QueryString "bitemyapp")
       let search = mkSearch (Just query) Nothing
       myTweet <- searchTweet search
       liftIO $
@@ -389,9 +435,8 @@ main = hspec $ do
       let sortSpec = DefaultSortSpec $ mkSort (FieldName "age") Ascending
       let search = Search Nothing
                    (Just IdentityFilter) (Just [sortSpec]) Nothing Nothing
-                   False (From 0) (Size 10)
-      reply <- searchByIndex testIndex search
-      let result = eitherDecode (responseBody reply) :: Either String (SearchResult Tweet)
+                   False (From 0) (Size 10) Nothing
+      result <- searchTweets search
       let myTweet = fmap (hitSource . head . hits . searchHits) result
       liftIO $
         myTweet `shouldBe` Right otherTweet
@@ -525,7 +570,6 @@ main = hspec $ do
       liftIO $
         myTweet `shouldBe` Right exampleTweet
 
-
     it "returns document for regexp filter" $ withTestEnv $ do
       _ <- insertData
       let filter = RegexpFilter (FieldName "user") (Regexp "bite.*app")
@@ -542,6 +586,20 @@ main = hspec $ do
                    (CacheName "test") False (CacheKey "key")
       let search = mkSearch Nothing (Just filter)
       searchExpectNoResults search
+
+    it "returns document for query filter, uncached" $ withTestEnv $ do
+      _ <- insertData
+      let filter = QueryFilter (TermQuery (Term "user" "bitemyapp") Nothing) True
+          search = mkSearch Nothing (Just filter)
+      myTweet <- searchTweet search
+      liftIO $ myTweet `shouldBe` Right exampleTweet
+
+    it "returns document for query filter, cached" $ withTestEnv $ do
+      _ <- insertData
+      let filter = QueryFilter (TermQuery (Term "user" "bitemyapp") Nothing) False
+          search = mkSearch Nothing (Just filter)
+      myTweet <- searchTweet search
+      liftIO $ myTweet `shouldBe` Right exampleTweet
 
   describe "Aggregation API" $ do
     it "returns term aggregation results" $ withTestEnv $ do
@@ -570,6 +628,32 @@ main = hspec $ do
       _ <- insertData
       searchTermsAggHint [GlobalOrdinals, GlobalOrdinalsHash, GlobalOrdinalsLowCardinality, Map]
 
+
+    it "can execute value_count aggregations" $ withTestEnv $ do
+      _ <- insertData
+      _ <- insertOther
+      let ags = mkAggregations "user_count" (ValueCountAgg (FieldValueCount (FieldName "user"))) <>
+                mkAggregations "bogus_count" (ValueCountAgg (FieldValueCount (FieldName "bogus")))
+      let search = mkAggregateSearch Nothing ags
+      let docCountPair k n = (k, object ["value" .= Number n])
+      res <- searchTweets search
+      liftIO $
+        fmap aggregations res `shouldBe` Right (Just (M.fromList [ docCountPair "user_count" 2
+                                                                 , docCountPair "bogus_count" 0
+                                                                 ]))
+
+    it "can execute filter aggregations" $ withTestEnv $ do
+      _ <- insertData
+      _ <- insertOther
+      let ags = mkAggregations "bitemyapps" (FilterAgg (FilterAggregation (TermFilter (Term "user" "bitemyapp") defaultCache) Nothing)) <>
+                mkAggregations "notmyapps" (FilterAgg (FilterAggregation (TermFilter (Term "user" "notmyapp") defaultCache) Nothing))
+      let search = mkAggregateSearch Nothing ags
+      let docCountPair k n = (k, object ["doc_count" .= Number n])
+      res <- searchTweets search
+      liftIO $
+        fmap aggregations res `shouldBe` Right (Just (M.fromList [ docCountPair "bitemyapps" 1
+                                                                 , docCountPair "notmyapps" 1
+                                                                 ]))
     -- Interaction of date serialization and date histogram aggregation is broken.
     -- it "returns date histogram aggregation results" $ withTestEnv $ do
     --   _ <- insertData
@@ -659,3 +743,22 @@ main = hspec $ do
   describe "Monoid (SearchHits a)" $ do
     prop "abides the monoid laws" $ eq $
       prop_Monoid (T :: T (SearchHits ()))
+
+  describe "mkDocVersion" $ do
+    prop "can never construct an out of range docVersion" $ \i ->
+      let res = mkDocVersion i
+      in case res of
+        Nothing -> property True
+        Just dv -> (dv >= minBound) .&&.
+                   (dv <= maxBound) .&&.
+                   docVersionNumber dv === i
+
+  describe "Enum DocVersion" $ do
+    it "follows the laws of Enum, Bounded" $ do
+      evaluate (succ maxBound :: DocVersion) `shouldThrow` anyErrorCall
+      evaluate (pred minBound :: DocVersion) `shouldThrow` anyErrorCall
+      evaluate (toEnum 0 :: DocVersion) `shouldThrow` anyErrorCall
+      evaluate (toEnum 9200000000000000001 :: DocVersion) `shouldThrow` anyErrorCall
+      enumFrom (pred maxBound :: DocVersion) `shouldBe` [pred maxBound, maxBound]
+      enumFrom (pred maxBound :: DocVersion) `shouldBe` [pred maxBound, maxBound]
+      enumFromThen minBound (pred maxBound :: DocVersion) `shouldBe` [minBound, pred maxBound]